+++ Develappers.BillomatNet/ClientService.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Diagnostics.CodeAnalysis;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
@@ -28,6 +27,7 @@ namespace Develappers.BillomatNet
         IEntityTagService<ClientTag, ClientTagFilter>
     {
         private readonly Configuration _configuration;
+        private const string EntityUrlFragment = "clients";
 
         /// <summary>
         /// Creates a new instance of <see cref="ClientService"/>.
@@ -55,7 +55,7 @@ namespace Develappers.BillomatNet
         public async Task<Account> MyselfAsync(CancellationToken token = default)
         {
             var httpClient = HttpClientFactory.Invoke();
-            var httpResponse = await httpClient.GetAsync(new Uri("/api/clients/myself", UriKind.Relative), token).ConfigureAwait(false);
+            var httpResponse = await httpClient.GetAsync(new Uri($"/api/{EntityUrlFragment}/myself", UriKind.Relative), token).ConfigureAwait(false);
             var jsonModel = JsonConvert.DeserializeObject<AccountWrapper>(httpResponse);
             return jsonModel.ToDomain();
         }
@@ -78,7 +78,7 @@ namespace Develappers.BillomatNet
         /// <returns>The client list or null if not found.</returns>
         public async Task<Types.PagedList<Client>> GetListAsync(Query<Client, ClientFilter> query, CancellationToken token = default)
         {
-            var jsonModel = await GetListAsync<ClientListWrapper>("/api/clients", QueryString.For(query), token).ConfigureAwait(false);
+            var jsonModel = await GetListAsync<ClientListWrapper>($"/api/{EntityUrlFragment}", QueryString.For(query), token).ConfigureAwait(false);
             return jsonModel.ToDomain();
         }
 
@@ -91,7 +91,7 @@ namespace Develappers.BillomatNet
         /// <exception cref="NotAuthorizedException">Thrown when the client is not accessible.</exception>
         public async Task<Client> GetByIdAsync(int id, CancellationToken token = default)
         {
-            var jsonModel = await GetItemByIdAsync<ClientWrapper>($"/api/clients/{id}", token).ConfigureAwait(false);
+            var jsonModel = await GetItemByIdAsync<ClientWrapper>($"/api/{EntityUrlFragment}/{id}", token).ConfigureAwait(false);
             return jsonModel.ToDomain();
         }
 
@@ -102,7 +102,7 @@ namespace Develappers.BillomatNet
                 throw new ArgumentException("invalid client id", nameof(id));
             }
 
-            return $"https://{_configuration.BillomatId}.billomat.net/app/clients/show/entityId/{id}";
+            return $"https://{_configuration.BillomatId}.billomat.net/app/{EntityUrlFragment}/show/entityId/{id}";
         }
 
         Task IEntityService<Client, ClientFilter>.DeleteAsync(int id, CancellationToken token)